Fix textMultiple done + textArea
