+++ lib/bloc/add_service_voucher/add_service_event.dart
@@ -11,10 +11,18 @@ class PostServiceVoucherEvent extends AddServiceVoucherEvent {
   final String bienSoXe;
 
   PostServiceVoucherEvent(this.sdt, this.bienSoXe);
+  @override
+  List<Object?> get props => [sdt, bienSoXe];
 }
 
 class SaveVoucherServiceEvent extends AddServiceVoucherEvent {
   final VoucherServiceRequest voucherServiceRequest;
+  final List<File>? listFile;
 
-  SaveVoucherServiceEvent(this.voucherServiceRequest);
+  SaveVoucherServiceEvent(
+    this.voucherServiceRequest,
+    this.listFile,
+  );
+  @override
+  List<Object?> get props => [voucherServiceRequest, listFile];
 }
