+++ lib/screens/menu/home/contract/update_contract.dart
@@ -27,8 +27,8 @@ import '../../../../models/widget_input_date.dart';
 import '../../../../src/models/model_generator/login_response.dart';
 import '../../../../src/pick_file_image.dart';
 import '../../../../storages/share_local.dart';
+import '../../../../widgets/multiple_widget.dart';
 import '../../../../widgets/widget_field_input_percent.dart';
-import '../customer/add_customer.dart';
 import '../customer/input_dropDown.dart';
 
 class EditContract extends StatefulWidget {
@@ -371,7 +371,7 @@ class _EditContractState extends State<EditContract> {
                                                                     : state.listEditData[index].data![index1].field_type == "HIDDEN"
                                                                         ? Container()
                                                                         : state.listEditData[index].data![index1].field_type == "TEXT_MULTI_NEW"
-                                                                            ? WidgetInputMulti(
+                                                                            ? InputMultipleWidget(
                                                                                 data: state.listEditData[index].data![index1],
                                                                                 onSelect: (data) {
                                                                                   addData[index].data[index1].value = data.join(",");
@@ -568,6 +568,8 @@ class _EditContractState extends State<EditContract> {
               padding: EdgeInsets.only(left: 10, top: 5, bottom: 5),
               child: Container(
                 child: TextFormField(
+                  minLines: data.field_type == 'TEXTAREA' ? 2 : 1,
+                  maxLines: data.field_type == 'TEXTAREA' ? 6 : 1,
                   style: TextStyle(fontSize: 14, fontWeight: FontWeight.w600),
                   keyboardType: data.field_type == "TEXT_NUMERIC"
                       ? TextInputType.number
