+++ lib/screens/menu/home/customer/form_edit.dart
@@ -34,6 +34,7 @@ import '../../../../src/src_index.dart';
 import 'package:multi_select_flutter/multi_select_flutter.dart';
 import '../../../../widgets/appbar_base.dart';
 import '../../../../widgets/loading_api.dart';
+import '../../../../widgets/multiple_widget.dart';
 import '../../../../widgets/widget_field_input_percent.dart';
 import '../../../add_service_voucher/add_service_voucher_step2_screen.dart';
 import 'input_dropDown.dart';
@@ -441,16 +442,10 @@ class _FormEditState extends State<FormEdit> {
                                                             : state.listEditData[index].data![index1].field_type == "HIDDEN"
                                                                 ? Container()
                                                                 : state.listEditData[index].data![index1].field_type == "TEXT_MULTI_NEW"
-                                                                    ? WidgetInputMulti(
+                                                                    ? InputMultipleWidget(
                                                                         data: state
                                                                             .listEditData[index]
                                                                             .data![index1],
-                                                                        onRemove:
-                                                                            (data) {
-                                                                          addData[index]
-                                                                              .data[index1]
-                                                                              .value = data.join(",");
-                                                                        },
                                                                         onSelect:
                                                                             (data) {
                                                                           addData[index]
@@ -645,6 +640,8 @@ class _FormEditState extends State<FormEdit> {
               padding: EdgeInsets.only(left: 10, top: 5, bottom: 5),
               child: Container(
                 child: TextFormField(
+                  minLines: data.field_type == 'TEXTAREA' ? 2 : 1,
+                  maxLines: data.field_type == 'TEXTAREA' ? 6 : 1,
                   style: TextStyle(fontSize: 14, fontWeight: FontWeight.w500),
                   keyboardType: data.field_special == "default"
                       ? TextInputType.text
@@ -829,188 +826,6 @@ class _FormEditState extends State<FormEdit> {
             files: AttackBloc.of(context).listFile));
       }
     }
-  }
-}
-
-class WidgetInputMulti extends StatefulWidget {
-  WidgetInputMulti(
-      {Key? key,
-      required this.data,
-      required this.onSelect,
-      required this.value,
-      required this.onRemove})
-      : super(key: key);
-
-  final CustomerIndividualItemData data;
-  Function onSelect;
-  Function onRemove;
-  List<String> value;
-
-  @override
-  State<WidgetInputMulti> createState() => _WidgetInputMultiState();
-}
-
-class _WidgetInputMultiState extends State<WidgetInputMulti> {
-  List<String> arr = [];
-  TextEditingController _editingController = TextEditingController();
-  bool check = false;
-  late FocusNode _focusNode;
-
-  @override
-  void initState() {
-    _focusNode = FocusNode();
-    arr = widget.value;
-    super.initState();
-  }
-
-  @override
-  void dispose() {
-    _focusNode.dispose();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      margin: EdgeInsets.only(bottom: 16),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          RichText(
-            textScaleFactor: MediaQuery.of(context).textScaleFactor,
-            text: TextSpan(
-              text: widget.data.field_label ?? '',
-              style: TextStyle(
-                  fontFamily: "Quicksand",
-                  fontSize: 14,
-                  fontWeight: FontWeight.w600,
-                  color: COLORS.BLACK),
-              children: <TextSpan>[
-                widget.data.field_require == 1
-                    ? TextSpan(
-                        text: '*',
-                        style: TextStyle(
-                            fontFamily: "Quicksand",
-                            fontSize: 14,
-                            fontWeight: FontWeight.w500,
-                            color: Colors.red))
-                    : TextSpan(),
-              ],
-            ),
-          ),
-          SizedBox(
-            height: 8,
-          ),
-          Container(
-            width: double.infinity,
-            decoration: BoxDecoration(
-                borderRadius: BorderRadius.circular(5),
-                border: Border.all(color: HexColor("#BEB4B4"))),
-            child: Padding(
-              padding: EdgeInsets.only(left: 10, top: 5, bottom: 5),
-              child: Focus(
-                onFocusChange: (status) {
-                  if (status == false) {
-                    if (_editingController.text != "") {
-                      arr.add(_editingController.text);
-                      widget.onSelect(arr);
-                      setState(() {
-                        check = !check;
-                      });
-                    }
-                    _editingController.text = "";
-                    _focusNode.unfocus();
-                  }
-                },
-                child: TextField(
-                  controller: _editingController,
-                  onEditingComplete: () {
-                    if (_editingController.text != "") {
-                      arr.add(_editingController.text);
-                      widget.onSelect(arr);
-                      setState(() {
-                        check = !check;
-                      });
-                    }
-                    _editingController.text = "";
-                    _focusNode.unfocus();
-                  },
-                  focusNode: _focusNode,
-                  style: TextStyle(fontSize: 14, fontWeight: FontWeight.w500),
-                  keyboardType: widget.data.field_special == "default"
-                      ? TextInputType.text
-                      : widget.data.field_special == "numberic"
-                          ? TextInputType.number
-                          : widget.data.field_special == "email-address"
-                              ? TextInputType.emailAddress
-                              : TextInputType.text,
-                  inputFormatters: [
-                    LengthLimitingTextInputFormatter(
-                        widget.data.field_maxlength != null
-                            ? int.parse(widget.data.field_maxlength!)
-                            : null),
-                  ],
-                  maxLengthEnforcement:
-                      MaxLengthEnforcement.truncateAfterCompositionEnds,
-                  decoration: InputDecoration(
-                      focusedBorder: InputBorder.none,
-                      enabledBorder: InputBorder.none,
-                      disabledBorder: InputBorder.none,
-                      isDense: true),
-                ),
-              ),
-            ),
-          ),
-          arr.length > 0
-              ? Container(
-                  margin: EdgeInsets.only(top: 8),
-                  child: Row(
-                    children: List.generate(
-                        arr.length,
-                        (index) => Container(
-                              margin: EdgeInsets.only(right: 8),
-                              child: Stack(
-                                clipBehavior: Clip.none,
-                                children: [
-                                  Container(
-                                      padding: EdgeInsets.only(
-                                          top: 3, bottom: 3, left: 8, right: 8),
-                                      decoration: BoxDecoration(
-                                          borderRadius:
-                                              BorderRadius.circular(100),
-                                          color: COLORS.BACKGROUND),
-                                      child: WidgetText(
-                                        title: arr[index],
-                                        style: AppStyle.DEFAULT_14,
-                                      )),
-                                  Positioned(
-                                    top: -13,
-                                    right: -8,
-                                    child: InkWell(
-                                        onTap: () {
-                                          arr.removeAt(index);
-                                          setState(() {
-                                            check = !check;
-                                          });
-                                        },
-                                        child: Container(
-                                          color: Colors.transparent,
-                                          padding: EdgeInsets.all(5),
-                                          child: WidgetText(
-                                            title: "x",
-                                            style: AppStyle.DEFAULT_16,
-                                          ),
-                                        )),
-                                  )
-                                ],
-                              ),
-                            )),
-                  ),
-                )
-              : Container()
-        ],
-      ),
-    );
   }
 }
